+++ src/routes/index.tsx
@@ -8,7 +8,5 @@ export {
     MediaRoute
 } from "./MediaRoute";
 
-export * from "./navigation";
-